Changes from SCU-1 to SSCU-1
These are the changes needed to get a simple test setup running with the
SSCU from BBH. The master-to-master sizes have changed from 8 bytes (uint64_t)
to 12 bytes (realized as array of uint8_t).

Also the Sync Manager Length therefore needs to be adjusted. The
commenting of cia_402_tx3 is due to the configuration of the
synapticon drive.

Signed-off-by: Matthias Schoepfer <msc@fp-robotics.com>